+++ controller/controller.go
@@ -22,10 +22,10 @@ type SessionData struct {
 	session *dg.Session
 	prefix  string
 
-	gifRequest *plugs.GifRequest
-	memeRecord *plugs.Record
-	pingRecord *plugs.Record
-	r34Request *plugs.Rule34Request
+	gifRequest  *plugs.GifRequest
+	memeRequest *plugs.MemeRequest
+	pingRecord  *plugs.Record
+	r34Request  *plugs.Rule34Request
 }
 
 // NewSessionData creates a SessionData
@@ -34,10 +34,10 @@ func NewSessionData(s *dg.Session) *SessionData {
 		session: s,
 		prefix:  `~`,
 
-		gifRequest: plugs.NewGifRequest(),
-		memeRecord: plugs.NewRecorder(),
-		pingRecord: plugs.NewRecorder(),
-		r34Request: plugs.NewRule34Request(),
+		gifRequest:  plugs.NewGifRequest(),
+		memeRequest: plugs.NewMemeRequest(),
+		pingRecord:  plugs.NewRecorder(),
+		r34Request:  plugs.NewRule34Request(),
 	}
 }
 
@@ -81,7 +81,7 @@ func (sd *SessionData) ExecuteTask(s *dg.Session, msg *dg.MessageCreate) {
 	case sd.prefix + "man":
 		res, err = plugs.Manual(req, s, msg)
 	case sd.prefix + "meme":
-		res, err = sd.memeRecord.RequestMeme(req, s, msg)
+		res, err = sd.memeRequest.RequestMeme(req, s, msg)
 	case sd.prefix + "ping":
 		res = sd.pingRecord.Pong(s, msg)
 	case sd.prefix + "rule34":